+++ R/image_glm.R
@@ -1,0 +1,3 @@
+#' @rdname persp
+#' @export
+image.glm <- image.lm