+++ R/image_loess.R
@@ -1,0 +1,3 @@
+#' @rdname persp
+#' @export
+image.loess <- image.lm