+++ R/persp_glm.R
@@ -1,0 +1,3 @@
+#' @rdname persp
+#' @export
+persp.glm <- persp.lm