+++ R/persp_loess.R
@@ -1,0 +1,3 @@
+#' @rdname persp
+#' @export
+persp.loess <- persp.lm